Well. Lots of stuff, work in progress.
This is a mess, lots of things are already known not to survive
to any finalized version, and this commit probably doesn't even
belong to this branch. However for reasons I haven't backed up
for some time and I don't want to go on with this any longer.
The branch is now development branch; once reasonably complete
set of features has been reached, the next phase will consist
of refactoring, cleaning up, writing tests and as these tasks
lead to finalization of parts of code, of commiting into more
civilized branch.